requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
biopython==1.84
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
datasets==3.1.0
dill==0.3.8
faiss-cpu==1.9.0.post1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
pyarrow==18.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
six==1.16.0
soupsieve==2.6
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tqdm==4.67.0
transformers==4.46.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.18.0
